% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{make_report}
\alias{make_report}
\title{Create a collections R Markdown report}
\usage{
make_report(path = "wnc_report.Rmd", open = interactive())
}
\arguments{
\item{path}{Where to write the new .Rmd file.}

\item{open}{Open the file in RStudio if possible.}
}
\description{
Copies the package's collections skeleton to a new .Rmd file.
}
