% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_wnc_report.R
\name{use_wnc_report}
\alias{use_wnc_report}
\title{Create a Wild Nematode Collections R Markdown report}
\usage{
use_wnc_report(path = "wnc_report.Rmd", open = interactive())
}
\arguments{
\item{path}{Where to write the new .Rmd file.}

\item{open}{Open the file in RStudio if possible.}
}
\description{
Copies the package's WNC report skeleton to a new .Rmd file.
}
